refactor: move lists and lists definitions models
