+++ vmngclient/endpoints/configuration_policy_data_prefix_list_builder.py
@@ -1,31 +1,8 @@
 # mypy: disable-error-code="empty-body"
-from ipaddress import IPv4Network
-from typing import List
-
-from pydantic import BaseModel, Field, validator
-
 from vmngclient.endpoints import APIEndpoints, delete, get, post, put
-from vmngclient.model.policy.policy_list import InfoTag, PolicyList, PolicyListId, PolicyListInfo, PolicyListPreview
+from vmngclient.model.policy.lists import DataPrefixList
+from vmngclient.model.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
 from vmngclient.typed_list import DataSequence
-
-
-class DataPrefixListEntry(BaseModel):
-    class Config:
-        allow_population_by_field_name = True
-
-    ip_prefix: str = Field(alias="ipPrefix", description="IP4 network prefixes separated by comma")
-
-    @validator("ip_prefix")
-    def check_network_prefixes(cls, ip_prefix: str):
-        nets = [IPv4Network(net.strip()) for net in ip_prefix.split(",")]
-        if len(nets) < 1:
-            raise ValueError("No network prefix provided")
-        return ip_prefix
-
-
-class DataPrefixList(PolicyList):
-    entries: List[DataPrefixListEntry]
-    type: str = Field(default="dataPrefix", const=True)
 
 
 class DataPrefixListEditPayload(DataPrefixList, PolicyListId):
