+++ vmngclient/model/policy/lists.py
@@ -1,0 +1,32 @@
+from typing import List, Literal, Optional, Union
+
+from pydantic import BaseModel, Field
+from typing_extensions import Annotated
+
+from vmngclient.model.policy.lists_entries import DataPrefixListEntry, SiteListEntry, VPNListEntry
+
+
+class PolicyListHeader(BaseModel):
+    name: str = Field(
+        regex="^[a-zA-Z0-9_-]{1,32}$",
+        description="Can include only alpha-numeric characters, hyphen '-' or underscore '_'; maximum 32 characters",
+    )
+    description: Optional[str] = "Desc Not Required"
+
+
+class DataPrefixList(PolicyListHeader):
+    type: Literal["dataPrefix"] = "dataPrefix"
+    entries: List[DataPrefixListEntry]
+
+
+class SiteList(PolicyListHeader):
+    type: Literal["site"] = "site"
+    entries: List[SiteListEntry]
+
+
+class VPNList(PolicyListHeader):
+    type: Literal["vpn"] = "vpn"
+    entries: List[VPNListEntry]
+
+
+AllPolicyLists = Annotated[Union[DataPrefixList, SiteList, VPNList], Field(discriminator="type")]
