+++ vmngclient/model/policy/policy_list.py
@@ -3,6 +3,10 @@ from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
+from vmngclient.model.policy.lists import AllPolicyLists
+
+PolicyList = AllPolicyLists
+
 
 class InfoTag(BaseModel):
     info_tag: Optional[str] = Field("", alias="infoTag")
@@ -10,16 +14,6 @@ class InfoTag(BaseModel):
 
 class PolicyListId(BaseModel):
     list_id: str = Field(alias="listId")
-
-
-class PolicyList(BaseModel):
-    name: str = Field(
-        regex="^[a-zA-Z0-9_-]{1,32}$",
-        description="Can include only alpha-numeric characters, hyphen '-' or underscore '_'; maximum 32 characters",
-    )
-    description: Optional[str] = "Desc Not Required"
-    type: str
-    entries: List
 
 
 class PolicyListInfo(PolicyListId, InfoTag):
@@ -30,10 +24,6 @@ class PolicyListInfo(PolicyListId, InfoTag):
     reference_count: int = Field(alias="referenceCount")
     references: List
     is_activated_by_vsmart: Optional[bool] = Field(None, alias="isActivatedByVsmart")
-
-
-class PolicyListEditPayload(PolicyList, PolicyListId):
-    pass
 
 
 class PolicyListPreview(BaseModel):
